+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## [0.3.5] - custom grouped checkbox 
+create customisable grouped checkbox 
+fix error 
 ## [0.3.2] - refactoring code 
 ## [0.3.1+3] - fix errors 
 ## [0.3.1+2] - remove direction groupedcheckbox